Moved all mineral operations into separate mineral.c file for increased readability and accessability
