+++ workspace/Fatherload/source/game.c
@@ -54,20 +54,6 @@ void start_game() {
 	update_sprite(gfx_horizontal, PLAYER_SPRITE_ID, 0,
 				  PLAYER_VPAL, 0, 0,
 				  player_x, player_y);
-/*	oamSet(&oamMain, 			// oam handler
-			PLAYER_SPRITE_ID, 	// Number of sprite
-			player_x, player_y, // Coordinates
-			0, 					// Priority
-			0, 					// Palette to use
-			SpriteSize_16x16, 	// Sprite size
-			SpriteColorFormat_16Color, // Color format
-			gfx_horizontal,	 	// Loaded graphic to display
-			-1, 				// Affine rotation to use (-1 none)
-			false, 				// Double size if rotating
-			false, 				// Hide this sprite
-			false, false,		// Horizontal or vertical flip
-			false 				// Mosaic
-	);*/
 	oamUpdate(&oamMain);
 	// set the starting backgrounds
 	init_main_bg();
@@ -129,20 +115,6 @@ void player_pressed_start() {
 		update_sprite(gfx_horizontal, PLAYER_SPRITE_ID, 1,
 					  PLAYER_VPAL, 0, 0,
 					  player_x, player_y);
-/*		oamSet(&oamMain,			// oam handler
-				PLAYER_SPRITE_ID, 	// Number of sprite
-				player_x, player_y, // Coordinates
-				0, 					// Priority
-				0, 					// Palette to use
-				SpriteSize_16x16, 	// Sprite size
-				SpriteColorFormat_16Color, // Color format
-				gfx_horizontal, 	// Loaded graphic to display
-				-1, 				// Affine rotation to use (-1 none)
-				false, 				// Double size if rotating
-				true, 				// Hide this sprite
-				false, false,		// Horizontal or vertical flip
-				false 				// Mosaic
-		);*/
 		oamUpdate(&oamMain);
 		swiDelay(11000000); // Delay to avoid going back out of start mode right after
 	} else {
@@ -150,20 +122,6 @@ void player_pressed_start() {
 		update_sprite(gfx_horizontal, PLAYER_SPRITE_ID, 0,
 					  PLAYER_HPAL, 0, 0,
 					  player_x, player_y);
-/*		oamSet(&oamMain, 			// oam handler
-				PLAYER_SPRITE_ID, 	// Number of sprite
-				player_x, player_y, // Coordinates
-				0, 					// Priority
-				0, 					// Palette to use
-				SpriteSize_16x16, 	// Sprite size
-				SpriteColorFormat_16Color, // Color format
-				gfx_horizontal, 	// Loaded graphic to display
-				-1, 				// Affine rotation to use (-1 none)
-				false, 				// Double size if rotating
-				false, 				// Hide this sprite
-				false, false, 		// Horizontal or vertical flip
-				false 				// Mosaic
-		);*/
 		oamUpdate(&oamMain);
 		release_start_display();
 		start_pressed = 0;
@@ -257,130 +215,3 @@ void update_state(){
 	oamUpdate(&oamMain);
 }
 
-void check_diamond(int position_x, int position_y){
-	int i;
-	for(i = 0; i < N_DIAMOND; i++){
-		if(!mineral[i].isDrilled && (
-		   (((position_y - mineral[i].y > 0) ? (position_y - mineral[i].y):(mineral[i].y - position_y)) < 16) // abs(diamond.y - position_y) < 16
-		&& (((position_x - mineral[i].x > 0) ? (position_x - mineral[i].x):(mineral[i].x - position_x)) < 16) // abs(diamond.x - position_x) < 16
-		)){
-			mineral_count++;
-			if(score_player < 999999){
-				score_player += SCORE_DIAMONDS;
-				if(score_player > 999999)
-					score_player = 999999;
-				score_changed = 1;
-			}
-			Audio_PlaySoundEX(SFX_COIN_PICKUP);
-			mineral[i].isDrilled = true;
-			update_sprite(diamond_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, DIAMOND_PAL, 0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-		}
-
-		if(((mineral[i].x > - 16) 		   && (mineral[i].x < screen_x + 256)) &&
-		   ((mineral[i].y > screen_y - 16) && (mineral[i].y < screen_y + 192))){
-			update_sprite(diamond_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, DIAMOND_PAL,0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-		}
-	}
-}
-
-void check_amazonite(int position_x, int position_y){
-	int i;
-	for(i = N_DIAMOND; i < N_DIAMOND + N_AMAZONITE; i++){
-		if(!mineral[i].isDrilled && (
-		(((position_y - mineral[i].y > 0) ? (position_y - mineral[i].y):(mineral[i].y - position_y)) < 16) //abs(diamond.y - position_y) < 16
-		&& (((position_x - mineral[i].x > 0) ? (position_x - mineral[i].x):(mineral[i].x - position_x)) < 16) // abs(diamond.x - position_x) < 16
-		)){
-			mineral_count++;
-			if(score_player < 999999){
-				score_player += SCORE_AMAZONITE;
-				if(score_player > 999999)
-					score_player = 999999;
-				score_changed = 1;
-			}
-			Audio_PlaySoundEX(SFX_COIN_PICKUP);
-			mineral[i].isDrilled = true;
-			//update_sprite(u16* gfx, int spriteID, int hide, int paletteNum, int x, int y)
-			update_sprite(amazonite_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, AMAZONITE_PAL, 0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-		}
-
-		if(((mineral[i].x >  - 16) && (mineral[i].x < screen_x + 256)) &&
-				((mineral[i].y > screen_y - 16) && (mineral[i].y < screen_y + 192))){
-
-			update_sprite(amazonite_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, AMAZONITE_PAL, 0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-		}
-	}
-}
-
-void check_bronze(int position_x, int position_y){
-	int i;
-	for(i = N_DIAMOND + N_AMAZONITE; i < N_DIAMOND + N_AMAZONITE + N_BRONZE; i++){
-		if(!mineral[i].isDrilled && (
-		(((position_y - mineral[i].y > 0) ? (position_y - mineral[i].y):(mineral[i].y - position_y)) < 16) //abs(diamond.y - position_y) < 16
-		&& (((position_x - mineral[i].x > 0) ? (position_x - mineral[i].x):(mineral[i].x - position_x)) < 16) // abs(diamond.x - position_x) < 16
-		)){
-			mineral_count++;
-			if(score_player < 999999){
-				score_player += SCORE_BRONZE;
-				if(score_player > 999999)
-					score_player = 999999;
-				score_changed = 1;
-			}
-			Audio_PlaySoundEX(SFX_COIN_PICKUP);
-			mineral[i].isDrilled = true;
-
-			update_sprite(bronze_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, BRONZE_PAL, 0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-
-		}
-
-		if(((mineral[i].x >  - 16) && (mineral[i].x < screen_x + 256)) &&
-				((mineral[i].y > screen_y - 16) && (mineral[i].y < screen_y + 192))){
-
-			update_sprite(bronze_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, BRONZE_PAL, 0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-		}
-	}
-}
-
-void check_alexxzandrite(int position_x, int position_y){
-	int i;
-	for(i = N_DIAMOND + N_AMAZONITE + N_BRONZE;
-		i < N_DIAMOND + N_AMAZONITE + N_BRONZE + N_ALEXXZANDRITE; i++){
-		if(!mineral[i].isDrilled && (
-		(((position_y - mineral[i].y > 0) ? (position_y - mineral[i].y):(mineral[i].y - position_y)) < 16) //abs(diamond.y - position_y) < 16
-		&& (((position_x - mineral[i].x > 0) ? (position_x - mineral[i].x):(mineral[i].x - position_x)) < 16) // abs(diamond.x - position_x) < 16
-		)){
-			mineral_count++;
-			if(score_player < 999999){
-				score_player += SCORE_ALEXANDRITE;
-				if(score_player > 999999)
-					score_player = 999999;
-				score_changed = 1;
-			}
-			Audio_PlaySoundEX(SFX_COIN_PICKUP);
-			mineral[i].isDrilled = true;
-
-			update_sprite(alexxzandrite_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, ALEXX_PAL, 0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-		}
-
-		if(((mineral[i].x >  - 16) && (mineral[i].x < screen_x + 256)) &&
-				((mineral[i].y > screen_y - 16) && (mineral[i].y < screen_y + 192))){
-
-			update_sprite(alexxzandrite_pic, OFFSET_MINERAL_SPRITE + i,
-						  mineral[i].isDrilled, ALEXX_PAL, 0, 0,
-						  mineral[i].x - screen_x, mineral[i].y - screen_y);
-		}
-	}
-}
