+++ workspace/Fatherload/source/graphics_main.h
@@ -25,10 +25,6 @@ void init_main_bg();
 
 void configureSprites();
 
-void initMinerals();
-
-
-
 void load_start_display();
 
 void print_start();
@@ -36,8 +32,6 @@ void print_start();
 void release_start_display();
 
 void restart_display();
-
-void hide_objects();
 
 void update_sprite(u16* gfx, int spriteID, int hide, int paletteNum, int hFlip, int vFlip, int x, int y);
