+++ workspace/Fatherload/source/mineral.h
@@ -1,0 +1,15 @@
+
+
+#include "fatherload_defines.h"
+
+void initMinerals();
+
+void hide_all_minerals();
+
+void check_alexxzandrite(int position_x, int position_y);
+
+void check_amazonite(int position_x, int position_y);
+
+void check_diamond(int position_x, int position_y);
+
+void check_bronze(int position_x, int position_y);
